superb-ai/requirements.txt
requests>=2.23.0
spb-cli>=0.5.2
opencv-python>=4.5.3.56
numpy>=1.16.5
argparse>=1.4.0